% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dem_old.R
\encoding{UTF-8}
\name{load_dem_old}
\alias{load_dem_old}
\title{Loads demographic information into R for demographics tables before 2022.}
\usage{
load_dem_old(
  file,
  merge_id = "EMPI",
  sep = ":",
  id_length = "standard",
  perc = 0.6,
  na = TRUE,
  identical = TRUE,
  nThread = parallel::detectCores() - 1,
  mrn_type = FALSE
)
}
\arguments{
\item{file}{string, full file path to Dem.txt.}

\item{merge_id}{string, column name to use to create \emph{ID_MERGE} column used to merge different datasets. Defaults to \emph{EPIC_PMRN},
as it is the preferred MRN in the RPDR system.}

\item{sep}{string, divider between hospital ID and MRN. Defaults to \emph{:}.}

\item{id_length}{string, indicating whether to modify MRN length based-on required values \emph{id_length = standard}, or to keep lengths as is \emph{id_length = asis}.
If \emph{id_length = standard} then in case of \emph{MGH, BWH, MCL, EMPI and PMRN} the length of the MRNs are corrected accordingly by adding zeros, or removing numeral from the beginning.
In other cases the lengths are unchanged. Defaults to \emph{standard}.}

\item{perc}{numeric, a number between 0-1 indicating which parsed ID columns to keep. Data present in \emph{perc x 100\%} of patients are kept.}

\item{na}{boolean, whether to remove columns with only NA values. Defaults to \emph{TRUE}.}

\item{identical}{boolean, whether to remove columns with identical values. Defaults to \emph{TRUE}.}

\item{nThread}{integer, number of threads to use to load data.}

\item{mrn_type}{boolean, should data in \emph{MRN_Type} and \emph{MRN} be parsed. Defaults to \emph{FALSE}, as it is not advised to parse these for all data sources as it takes considerable time.}
}
\value{
data table, with demographic information data.
\describe{
 \item{ID_MERGE}{numeric, defined IDs by \emph{merge_id}, used for merging later.}
 \item{ID_dem_EMPI}{string, Unique Partners-wide identifier assigned to the patient used to consolidate patient information.
 from \emph{dem} datasource, corresponds to EMPI in RPDR. Data is formatted using pretty_mrn().}
 \item{ID_dem_PMRN}{string, Epic medical record number. This value is unique across Epic instances within the Partners network.
 from \emph{dem} datasource, corresponds to EPIC_PMRN in RPDR. Data is formatted using pretty_mrn().}
 \item{ID_dem_loc}{string, if mrn_type == TRUE, then the data in \emph{MRN_Type} and \emph{MRN} are parsed into IDs corresponding to locations \emph{(loc)}. Data is formatted using pretty_mrn().}
 \item{gender}{string, Patient's legal sex, corresponds to Gender in RPDR.}
 \item{time_date_of_birth}{POSIXct, Patient's date of birth, corresponds to Date_of_Birth in RPDR. Converted to POSIXct format.}
 \item{age}{string, Patient's current age (or age at death), corresponds to Age in RPDR.}
 \item{language}{string, Patient's preferred spoken language, corresponds to Language in RPDR.}
 \item{race}{string, Patient's primary race, corresponds to Race in RPDR.}
 \item{marital}{string, Patient's current marital status, corresponds to Marital_Status in RPDR.}
 \item{religion}{string, Patient-identified religious preference, corresponds to Religion in RPDR.}
 \item{veteran}{string, Patient's current military veteran status, corresponds to Is_a_veteran in RPDR.}
 \item{country_dem}{string, Patient's current country of residence from dem datasource, corresponds to Country in RPDR.}
 \item{zip_dem}{string, Mailing zip code of patient's primary residence from dem datasource, corresponds to Zip_code in RPDR.Formatted to 5 character zip codes.}
 \item{vital_status}{string, Identifies if the patient is living or deceased.
 This data is updated monthly from the Partners registration system and the Social Security Death Master Index, corresponds to Vital_Status in RPDR. Punctuation marks are removed.}
 \item{time_date_of_death}{POSIXct, Recorded date of death from source in 'Vital_Status'.
 Date of death information obtained solely from the Social Security Death Index will not be reported until 3 years after death due to privacy concerns.
 If the value is independently documented by a Partners entity within the 3 year window then the date will be displayed. corresponds to Date_of_Death in RPDR. Converted to POSIXct format.}
}
}
\description{
Loads patient demographic  and vital status information into the R environment. Since version 0.2.2 of the software, this function supports the old demographics table data definitions
and is identical to the \emph{load_dem} function of previous versions of the software.
}
\examples{
\dontrun{
#Using defaults
d_dem <- load_dem_old(file = "test_Dem.txt")

#Use sequential processing
d_dem <- load_dem_old(file = "test_Dem.txt", nThread = 1)

#Use parallel processing and parse data in MRN_Type and MRN columns and keep all IDs
d_dem <- load_dem_old(file = "test_Dem.txt", nThread = 20, mrn_type = TRUE, perc = 1)
}
}
