% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rad.R
\encoding{UTF-8}
\name{load_rad}
\alias{load_rad}
\title{Loads radiology notes into R.}
\usage{
load_rad(
  file,
  merge_id = "EMPI",
  sep = ":",
  id_length = "standard",
  perc = 0.6,
  na = TRUE,
  identical = TRUE,
  nThread = 4,
  mrn_type = FALSE
)
}
\arguments{
\item{file}{string, full file path to Rad.txt.}

\item{merge_id}{string, column name to use to create \emph{ID_MERGE} column used to merge different datasets. Defaults to \emph{EPIC_PMRN},
as it is the preferred MRN in the RPDR system.}

\item{sep}{string, divider between hospital ID and MRN. Defaults to \emph{:}.}

\item{id_length}{string, indicating whether to modify MRN length based-on required values \emph{id_length = standard}, or to keep lengths as is \emph{id_length = asis}.
If \emph{id_length = standard} then in case of \emph{MGH, BWH, MCL, EMPI and PMRN} the length of the MRNs are corrected accordingly by adding zeros, or removing numeral from the beginning.
In other cases the lengths are unchanged. Defaults to \emph{standard}.}

\item{perc}{numeric, a number between 0-1 indicating which parsed ID columns to keep. Data present in \emph{perc x 100\%} of patients are kept.}

\item{na}{boolean, whether to remove columns with only NA values. Defaults to \emph{TRUE}.}

\item{identical}{boolean, whether to remove columns with identical values. Defaults to \emph{TRUE}.}

\item{nThread}{integer, number of threds to use by data.table for reading data.}

\item{mrn_type}{boolean, should data in \emph{MRN_Type} and \emph{MRN} be parsed. Defaults to \emph{FALSE}, as it is not advised to parse these for all data sources as it takes considerable time.}
}
\value{
data table, with radiological notes information.
\describe{
 \item{ID_MERGE}{numeric, defined IDs by \emph{merge_id}, used for merging later.}
 \item{ID_rad_EMPI}{string, Unique Partners-wide identifier assigned to the patient used to consolidate patient information
 from \emph{rad} datasource, corresponds to EMPI in RPDR. Data is formatted using pretty_mrn().}
 \item{ID_rad_PMRN}{string, Epic medical record number. This value is unique across Epic instances within the Partners network
 from \emph{rad} datasource, corresponds to EPIC_PMRN in RPDR. Data is formatted using pretty_mrn().}
 \item{ID_rad_loc}{string, if mrn_type == TRUE, then the data in \emph{MRN_Type} and \emph{MRN} are parsed into IDs corresponding to locations \emph{(loc)}. Data is formatted using pretty_mrn().}
 \item{time_rad_exam}{POSIXct, Date when the report was filed, corresponds to Report_Date_Time in RPDR. Converted to POSIXct format.}
 \item{rad_rep_num}{string, Source-specific identifier used to reference the report, corresponds to Report_Number in RPDR.}
 \item{rad_rep_desc}{string, Type of procedure detailed in the report, corresponds to Report_Description in RPDR.}
 \item{rad_rep_status}{string, Completion status of the note/report, corresponds to Report_Status in RPDR. }
 \item{rad_rep_type}{string, This will always default to RAD, corresponds to Report_Type in RPDR.}
 \item{rad_rep_txt}{string, Full narrative text contained in the note/report, corresponds to Report_Text in RPDR.}
 }
}
\description{
Loads radiology notes information into the R environment.
}
\examples{
\dontrun{
#Using defaults
d_rad <- load_rad(file = "test_Rad.txt")

#Use sequential processing
d_rad <- load_rad(file = "test_Rad.txt", nThread = 1)

#Use parallel processing and parse data in MRN_Type and MRN columns and keep all IDs
d_rad <- load_rad(file = "test_Rad.txt", nThread = 20, mrn_type = TRUE, perc = 1)
}
}
