% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rdt.R
\encoding{UTF-8}
\name{load_rdt}
\alias{load_rdt}
\title{Loads radiology procedures data into R.}
\usage{
load_rdt(
  file,
  merge_id = "EMPI",
  sep = ":",
  id_length = "standard",
  perc = 0.6,
  na = TRUE,
  identical = TRUE,
  nThread = parallel::detectCores() - 1,
  mrn_type = FALSE
)
}
\arguments{
\item{file}{string, full file path to Rdt.txt.}

\item{merge_id}{string, column name to use to create \emph{ID_MERGE} column used to merge different datasets. Defaults to \emph{EPIC_PMRN},
as it is the preferred MRN in the RPDR system.}

\item{sep}{string, divider between hospital ID and MRN. Defaults to \emph{:}.}

\item{id_length}{string, indicating whether to modify MRN length based-on required values \emph{id_length = standard}, or to keep lengths as is \emph{id_length = asis}.
If \emph{id_length = standard} then in case of \emph{MGH, BWH, MCL, EMPI and PMRN} the length of the MRNs are corrected accordingly by adding zeros, or removing numeral from the beginning.
In other cases the lengths are unchanged. Defaults to \emph{standard}.}

\item{perc}{numeric, a number between 0-1 indicating which parsed ID columns to keep. Data present in \emph{perc x 100\%} of patients are kept.}

\item{na}{boolean, whether to remove columns with only NA values. Defaults to \emph{TRUE}.}

\item{identical}{boolean, whether to remove columns with identical values. Defaults to \emph{TRUE}.}

\item{nThread}{integer, number of threads to use to load data.}

\item{mrn_type}{boolean, should data in \emph{MRN_Type} and \emph{MRN} be parsed. Defaults to \emph{FALSE}, as it is not advised to parse these for all data sources as it takes considerable time.}
}
\value{
data table, with radiological exam information.
\describe{
 \item{ID_MERGE}{numeric, defined IDs by \emph{merge_id}, used for merging later.}
 \item{ID_rdt_EMPI}{string, Unique Partners-wide identifier assigned to the patient used to consolidate patient information
 from \emph{rdt} datasource, corresponds to EMPI in RPDR. Data is formatted using pretty_mrn().}
 \item{ID_rdt_PMRN}{string, Epic medical record number. This value is unique across Epic instances within the Partners network
 from \emph{rdt} datasource, corresponds to EPIC_PMRN in RPDR. Data is formatted using pretty_mrn().}
 \item{ID_rdt_loc}{string, if mrn_type == TRUE, then the data in \emph{MRN_Type} and \emph{MRN} are parsed into IDs corresponding to locations \emph{(loc)}. Data is formatted using pretty_mrn().}
 \item{time_rdt_exam}{POSIXct, Date of the radiology exam, corresponds to Date in RPDR. Converted to POSIXct format.}
 \item{rdt_mode}{string, Modality of the exam, corresponds to Mode in RPDR.}
 \item{rdt_group}{string, Higher-level grouping concept used to consolidate similar procedures across hospitals, corresponds to Group in RPDR.}
 \item{rdt_test_code}{string, Internal identifier for the procedure used by the source system, corresponds to Test_Code in RPDR.}
 \item{rdt_test_desc}{string, Full name of the exam/study performed, corresponds to Test_Description in RPDR.}
 \item{rdt_accession}{string, Identifier assigned to the report or procedure for Radiology tracking purposes, corresponds to Accession_Number in RPDR.}
 \item{rdt_provider}{string, Ordering or authorizing provider for the study, corresponds to Provider in RPDR.}
 \item{rdt_clinic}{string, Specific department/location where the procedure was ordered or performed, corresponds to Clinic in RPDR.}
 \item{rdt_hosp}{string, Facility where the order was entered, corresponds to Hospital in RPDR.}
 \item{rdt_inpatient}{string, Classifies the type of encounter where the procedure was performed, corresponds to Inpatient_Outpatient in RPDR.}
 }
}
\description{
Loads radiology procedures information into the R environment.
}
\examples{
\dontrun{
#Using defaults
d_rdt <- load_rdt(file = "test_Rdt.txt")

#Use sequential processing
d_rdt <- load_rdt(file = "test_Rdt.txt", nThread = 1)

#Use parallel processing and parse data in MRN_Type and MRN columns and keep all IDs
d_rdt <- load_rdt(file = "test_Rdt.txt", nThread = 20, mrn_type = TRUE, perc = 1)
}
}
