% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_numbers.R
\encoding{UTF-8}
\name{pretty_numbers}
\alias{pretty_numbers}
\title{Converts numerical codes to universal format specified by length.}
\usage{
pretty_numbers(v, length_final = 5, remove_from_back = 4)
}
\arguments{
\item{v}{vector, integer or sting vector with numerical values.}

\item{length_final}{numeric, the length of the final string. Defaults to \emph{5} for zip code conversions.}

\item{remove_from_back}{numeric, the number of digits to remove from the back of the string. If \emph{NULL}, then removes characters from back more than specified in \emph{length_final}.
Defaults to \emph{4} for zip code conversions by removing the add-on codes.}
}
\value{
vector, with characters formatted accordingly.
}
\description{
Creates numerical strings with given lengths by removing additional characters from the back and adding leading zeros if necessary.
}
