% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_bar.R
\encoding{UTF-8}
\name{progress_bar}
\alias{progress_bar}
\title{Provides progress bar for foreach loops.}
\usage{
progress_bar(n, type)
}
\arguments{
\item{n}{integer, number of iterations}
}
\value{
progress bar
}
\description{
Provides progress bar based-on iteration cycles in foreach loops. Does not work for some reason.
In the call of foreach add: .combine=progress_bar(n = nThread, type = "c")
https://gist.github.com/kvasilopoulos/d49499ea854541924a8a4cc43a77fed0
}
\keyword{internal}
