% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_column.R
\encoding{UTF-8}
\name{remove_column}
\alias{remove_column}
\title{Delete columns with all NA or all identical data.}
\usage{
remove_column(dt, na = TRUE, identical = TRUE)
}
\arguments{
\item{dt}{data.table, to manipulate.}

\item{na}{boolean, to delete columns where all data elements are NA.}

\item{identical}{boolean, to delete columns where all data elements are the same.}
}
\value{
data table, with data.
}
\description{
Delete columns where all data elements are NA or the same.
}
\keyword{internal}
