% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_text.R
\encoding{UTF-8}
\name{pretty_text}
\alias{pretty_text}
\title{Removes spaces, special characters and capitals from string vector.}
\usage{
pretty_text(
  v,
  remove_after = FALSE,
  remove_punc = FALSE,
  remove_white = FALSE,
  add_na = TRUE
)
}
\arguments{
\item{v}{vector, integer or sting vector with numerical values.}

\item{remove_after}{boolean whether to remove text after \emph{-}. Defaults to \emph{FALSE}.}

\item{remove_punc}{boolean, whether to remove punctuation marks. Defaults to \emph{FALSE}.}

\item{remove_white}{boolean, whether to remove white spaces. Defaults to \emph{FALSE}.}

\item{add_na}{boolean, whether to change text indicating NA to NA values in R. Defaults to \emph{TRUE}.}
}
\value{
vector, with characters formatted accordingly.
}
\description{
Removes paces, special characters and capitals from string vector and converts unknowns to NA.
}
